Fix CSRF by not checking the value of its cookie
The idea of csrf is that you need to POST the token apart from the cookie, which will be sent when making a request from another page via JS (if that's allowed).
So the CSRF token must be sent either through a form or with JS via the X-Csrf-Token header.